requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==4.2.9
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
sentry-sdk==1.39.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
